@@ -20,52 +20,30 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>sk.lazyman</groupId>
-    <artifactId>gizmo-v3</artifactId>
-    <version>1.0-SNAPSHOT</version>
-    <packaging>war</packaging>
+    <groupId>com.evolveum</groupId>
+    <artifactId>gizmo-v4</artifactId>
+    <version>4.0-SNAPSHOT</version>
+    <packaging>jar</packaging>
 
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.4.0</version>
+        <version>3.1.5</version>
     </parent>
 
     <properties>
-        <java.version>11</java.version>
+        <java.version>17</java.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <apt-maven-plugin.version>1.1.3</apt-maven-plugin.version>
         <testng.version>6.8.8</testng.version>
-        <wicket.version>8.7.0</wicket.version>
-        <wicket.bootstrap.version>2.0.13</wicket.bootstrap.version>
-        <apache.poi.version>4.1.2</apache.poi.version>
+        <wicket.version>10.0.0-M2</wicket.version>
+<!--        <wicket.bootstrap.version>2.0.13</wicket.bootstrap.version>-->
+        <apache.poi.version>5.2.2</apache.poi.version>
+        <spring-boot.version>3.1.5</spring-boot.version>
+        <wicket.chartjs.version>0.3</wicket.chartjs.version>
     </properties>
 
     <dependencies>
-
-        <dependency>
-            <groupId>org.webjars</groupId>
-            <artifactId>AdminLTE</artifactId>
-            <version>3.0.5</version>
-        </dependency>
-        <dependency>
-            <groupId>org.webjars</groupId>
-            <artifactId>bootstrap</artifactId>
-            <version>4.4.1</version>
-        </dependency>
-        <dependency>
-            <groupId>org.webjars</groupId>
-            <artifactId>font-awesome</artifactId>
-            <version>4.7.0</version>
-        </dependency>
-
-
-<!--        <dependency>-->
-<!--            <groupId>org.webjars.bower</groupId>-->
-<!--            <artifactId>bootstrap-multiselect</artifactId>-->
-<!--            <version>0.9.15</version>-->
-<!--        </dependency>-->
-        <!-- SPRING -->
         <dependency>
             <groupId>org.springframework</groupId>
             <artifactId>spring-core</artifactId>
