@@ -18,10 +18,9 @@ package sk.lazyman.gizmo.data;
 
 import sk.lazyman.gizmo.util.GizmoUtils;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 import java.io.Serializable;
 import java.time.LocalDate;
-import java.util.Date;
 import java.util.Set;
 
 /**
