@@ -19,7 +19,6 @@ package sk.lazyman.gizmo.component.form;
 import org.apache.wicket.AttributeModifier;
 import org.apache.wicket.markup.html.WebMarkupContainer;
 import org.apache.wicket.markup.html.form.FormComponent;
-import org.apache.wicket.model.AbstractReadOnlyModel;
 import org.apache.wicket.model.IModel;
 
 /**
@@ -30,7 +29,7 @@ public class FormUtils {
     public static WebMarkupContainer createFormGroup(String id, final FormComponent formComponent) {
         WebMarkupContainer formGroup = new WebMarkupContainer(id);
         formGroup.add(formComponent);
-        formGroup.add(AttributeModifier.append("class", new AbstractReadOnlyModel<String>() {
+        formGroup.add(AttributeModifier.append("class", new IModel<String>() {
 
             @Override
             public String getObject() {
