@@ -16,18 +16,22 @@
 
 package sk.lazyman.gizmo.web;
 
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpSession;
 import org.apache.wicket.ajax.AjaxRequestTarget;
-import org.apache.wicket.ajax.markup.html.form.AjaxSubmitLink;
 import org.apache.wicket.markup.head.IHeaderResponse;
-import org.apache.wicket.markup.html.form.Form;
-import org.apache.wicket.markup.html.form.PasswordTextField;
-import org.apache.wicket.markup.html.form.RequiredTextField;
 import org.apache.wicket.model.IModel;
 import org.apache.wicket.model.Model;
-import org.apache.wicket.model.PropertyModel;
+import org.apache.wicket.protocol.http.WebApplication;
+import org.apache.wicket.protocol.http.servlet.ServletWebRequest;
+import org.apache.wicket.request.cycle.RequestCycle;
+import org.springframework.security.authentication.AnonymousAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.web.WebAttributes;
 import org.wicketstuff.annotation.mount.MountPath;
 import sk.lazyman.gizmo.component.MainFeedback;
-import sk.lazyman.gizmo.component.form.FormUtils;
+import sk.lazyman.gizmo.security.GizmoApplication;
 import sk.lazyman.gizmo.security.GizmoAuthWebSession;
 import sk.lazyman.gizmo.web.app.PageDashboard;
 
@@ -48,6 +52,11 @@ public class PageLogin extends PageTemplate {
     private IModel<LoginDto> model = new Model(new LoginDto());
 
     public PageLogin() {
+        if (getPrincipal() != null) {
+            WebApplication app = GizmoApplication.get();
+            setResponsePage(app.getHomePage());
+        }
+
         initLayout();
     }
 
@@ -57,48 +66,81 @@ public class PageLogin extends PageTemplate {
 //        response.render(CssHeaderItem.forReference(new LessResourceReference(PageLogin.class, "PageLogin.less")));
     }
 
+    private Object getPrincipal() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication == null) {
+            return null;
+        }
+        if (authentication instanceof AnonymousAuthenticationToken) {
+            return null;
+        }
+        return authentication.getPrincipal();
+    }
+
     private void initLayout() {
+
         MainFeedback feedback = new MainFeedback(ID_FEEDBACK);
         feedback.setOutputMarkupId(true);
         add(feedback);
 
-        Form<?> form = new Form<>(ID_FORM) {
-
-            @Override
-            protected void onSubmit() {
-                GizmoAuthWebSession session = GizmoAuthWebSession.getSession();
-
-                LoginDto dto = model.getObject();
-                if (session.authenticate(dto.getUsername(), dto.getPassword())) {
-                    setResponsePage(PageDashboard.class);
-                }
-            }
-        };
-        add(form);
-
-        RequiredTextField username = new RequiredTextField(ID_USERNAME, new PropertyModel(model, LoginDto.F_USERNAME));
-        FormUtils.addPlaceholderAndLabel(username, createStringResource("PageLogin.username"));
-        form.add(username);//FormUtils.createFormGroup(ID_USERNAME_GROUP, username));
-
-        PasswordTextField password = new PasswordTextField(ID_PASSWORD, new PropertyModel(model, LoginDto.F_PASSWORD));
-        FormUtils.addPlaceholderAndLabel(password, createStringResource("PageLogin.password"));
-        form.add(password);//FormUtils.createFormGroup(ID_PASSWORD_GROUP, password));
 
 
+//        Form<?> form = new Form<>(ID_FORM) {
+//
+//            @Override
+//            protected void onSubmit() {
+//                GizmoAuthWebSession session = GizmoAuthWebSession.getSession();
+//
+//                LoginDto dto = model.getObject();
+//                if (session.authenticate(dto.getUsername(), dto.getPassword())) {
+//                    setResponsePage(PageDashboard.class);
+//                }
+//            }
+//        };
+//        add(form);
 
-        AjaxSubmitLink signin = new AjaxSubmitLink(ID_BTN_SIGNIN) {
+//        RequiredTextField username = new RequiredTextField(ID_USERNAME, new PropertyModel(model, LoginDto.F_USERNAME));
+//        FormUtils.addPlaceholderAndLabel(username, createStringResource("PageLogin.username"));
+//        form.add(username);//FormUtils.createFormGroup(ID_USERNAME_GROUP, username));
+//
+//        PasswordTextField password = new PasswordTextField(ID_PASSWORD, new PropertyModel(model, LoginDto.F_PASSWORD));
+//        FormUtils.addPlaceholderAndLabel(password, createStringResource("PageLogin.password"));
+//        form.add(password);//FormUtils.createFormGroup(ID_PASSWORD_GROUP, password));
+//
 
-            @Override
-            protected void onError(AjaxRequestTarget target) {
-                target.add(PageLogin.this.get(ID_FEEDBACK));
-            }
 
-            @Override
-            protected void onSubmit(AjaxRequestTarget target) {
-                loginPerformed(target);
-            }
-        };
-        form.add(signin);
+//        AjaxSubmitLink signin = new AjaxSubmitLink(ID_BTN_SIGNIN) {
+//
+//            @Override
+//            protected void onError(AjaxRequestTarget target) {
+//                target.add(PageLogin.this.get(ID_FEEDBACK));
+//            }
+//
+//            @Override
+//            protected void onSubmit(AjaxRequestTarget target) {
+//                loginPerformed(target);
+//            }
+//        };
+//        form.add(signin);
+    }
+
+    @Override
+    protected void onConfigure() {
+        super.onConfigure();
+
+        ServletWebRequest req = (ServletWebRequest) RequestCycle.get().getRequest();
+        HttpServletRequest httpReq = req.getContainerRequest();
+        HttpSession httpSession = httpReq.getSession();
+
+        Exception ex = (Exception) httpSession.getAttribute(WebAttributes.AUTHENTICATION_EXCEPTION);
+        if (ex == null) {
+            return;
+        }
+
+        String key = ex.getMessage() != null ? ex.getMessage() : "web.security.provider.unavailable";
+        error(getString(key));
+
+        httpSession.removeAttribute(WebAttributes.AUTHENTICATION_EXCEPTION);
     }
 
     private void loginPerformed(AjaxRequestTarget target) {
