@@ -19,7 +19,6 @@ package sk.lazyman.gizmo.web.app;
 import org.apache.wicket.ajax.AjaxRequestTarget;
 import org.apache.wicket.extensions.markup.html.repeater.data.table.IColumn;
 import org.apache.wicket.extensions.markup.html.repeater.data.table.PropertyColumn;
-import org.apache.wicket.model.AbstractReadOnlyModel;
 import org.apache.wicket.model.IModel;
 import org.apache.wicket.request.mapper.parameter.PageParameters;
 import org.springframework.data.domain.Sort;
@@ -63,7 +62,7 @@ public class PageUsers extends PageAppUsers {
 
             @Override
             protected IModel<String> createIconModel(final IModel<User> rowModel) {
-                return new AbstractReadOnlyModel<String>() {
+                return new IModel<String>() {
 
                     @Override
                     public String getObject() {
